requirements.txt
requests>=2,<3
msal>=0.6.1,<2
python-docx
python-dotenv
azure-search-documents
azure-ai-formrecognizer
langchain
tiktoken
PyPDF2
openai[azure]
tenacity
azure-identity
markitdown
httpx==0.27.2 